Added dtype support

+++ optimized_method/optimized_svd.py
@@ -1,9 +1,9 @@
 import numpy as np
 from metrics.metrics import _extract_diagonal
 
-def optimized_svd(A):
+def optimized_svd(A, dtype=np.float64):
     U, S, Vt = np.linalg.svd(A, full_matrices=False)
-    return U, S, Vt
+    return U.astype(dtype), S.astype(dtype), Vt.astype(dtype)
 
 def reconstruct(U, S, Vt, k=None):
     if k is not None:
